feat: support markdown syntax in ai writer block
