+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/plugins.dart
@@ -24,9 +24,11 @@ export 'header/cover_editor_bloc.dart';
 export 'header/custom_cover_picker.dart';
 export 'header/document_cover_widget.dart';
 export 'heading/heading_toolbar_item.dart';
+export 'image/custom_image_block_component/custom_image_block_component.dart';
 export 'image/custom_image_block_component/image_menu.dart';
 export 'image/image_selection_menu.dart';
 export 'image/mobile_image_toolbar_item.dart';
+export 'image/multi_image_block_component/multi_image_block_component.dart';
 export 'image/multi_image_block_component/multi_image_menu.dart';
 export 'inline_math_equation/inline_math_equation.dart';
 export 'inline_math_equation/inline_math_equation_toolbar_item.dart';
@@ -36,6 +38,7 @@ export 'link_preview/link_preview_cache.dart';
 export 'link_preview/link_preview_menu.dart';
 export 'math_equation/math_equation_block_component.dart';
 export 'math_equation/mobile_math_equation_toolbar_item.dart';
+export 'mention/mention_block.dart';
 export 'mobile_floating_toolbar/custom_mobile_floating_toolbar.dart';
 export 'mobile_toolbar_v3/aa_toolbar_item.dart';
 export 'mobile_toolbar_v3/add_block_toolbar_item.dart';
@@ -49,7 +52,7 @@ export 'mobile_toolbar_v3/toolbar_item_builder.dart';
 export 'mobile_toolbar_v3/undo_redo_toolbar_item.dart';
 export 'mobile_toolbar_v3/util.dart';
 export 'numbered_list/numbered_list_icon.dart';
-export 'openai/widgets/auto_completion_node_widget.dart';
+export 'openai/widgets/ai_writer_block_component.dart';
 export 'openai/widgets/smart_edit_node_widget.dart';
 export 'openai/widgets/smart_edit_toolbar_item.dart';
 export 'outline/outline_block_component.dart';
@@ -58,12 +61,9 @@ export 'quote/quote_block_shortcuts.dart';
 export 'shortcuts/character_shortcuts.dart';
 export 'shortcuts/command_shortcuts.dart';
 export 'slash_menu/slash_menu_items.dart';
+export 'sub_page/sub_page_block_component.dart';
 export 'table/table_menu.dart';
 export 'table/table_option_action.dart';
 export 'todo_list/todo_list_icon.dart';
 export 'toggle/toggle_block_component.dart';
 export 'toggle/toggle_block_shortcuts.dart';
-export 'sub_page/sub_page_block_component.dart';
-export 'mention/mention_block.dart';
-export 'image/custom_image_block_component/custom_image_block_component.dart';
-export 'image/multi_image_block_component/multi_image_block_component.dart';
